+++ resources/views/seeds/table.blade.php
@@ -28,15 +28,21 @@
             </td>
             <td>
                 <div class="btn-group" role="group" aria-label="Basic example">
+
+                    @can('varieties-edit')
                  <a data-toggle="tooltip" data-placement="top" title="Edit"
                href="{{ route('backend.seeds.edit',$crop->id)}}" class="btn btn-sm  btn-outline-success">
                     <i class="fa fa-edit"></i>
                 </a>
+                @endcan
+                @can('varieties-delete')
                   <a href="javascript:;" data-toggle="modal" onclick="deleteData({{$crop->id}})"
               data-target="#DeleteModal" class="btn   btn-sm  btn-outline-danger">
               DELETE
               <i class="fa fa-trash-o" data-toggle="tooltip" data-placement="top" title="Delete"></i>
-              </a></div>
+              </a>
+              @endcan
+            </div>
             </td>
             @endforeach
         </tr>
