+++ resources/views/users/table.blade.php
@@ -1,4 +1,4 @@
-<table id="Data_Table" class="table table-bordered display nowrap" style="width:100%">
+<table id="Data_Table" class="table table-bordered">
     <thead>
     <tr>
         <th>Action</th>
@@ -12,24 +12,29 @@
     @foreach ($users as $key => $user)
         <tr>
             <td>
+                 @can('users-list')
                 <a data-toggle="tooltip" data-placement="top" title="View Details"
                 class="btn  btn-sm btn-outline-info" href="{{ route('backend.users.show',$user->id) }}">
                     <i class="fa fa-search"></i>
                 </a>
+                 @endcan
+                  @can('users-edit')
                 <a data-toggle="tooltip" data-placement="top" title="Edit" href="{{ route('backend.users.edit',$user->id)}}" class="btn btn-sm  btn-outline-success">
                     <i class="fa fa-edit"></i>
                 </a>
+                  @endcan
                 <!-- /.btn  btn-default -->
                 @if($user->id == config('custom.default_user_id') || $user->id ==  Auth::user()->id)
                     <button onclick="return false" type="submit" class="btn  btn-sm  btn-danger disabled">
                         <i class="fa fa-times"></i>
                     </button>
                 @else
+             @can('users-delete')
                     <a href="javascript:;" data-toggle="modal" onclick="deleteData({{$user->id}})"
               data-target="#DeleteModal" class="btn   btn-sm  btn-outline-danger">
               <i class="fa fa-trash" data-toggle="tooltip" data-placement="top" title="Delete"></i>
-
-            </a>
+                </a>
+              @endcan
                 @endif
 
             </td>
